+++ src/hooks/useEmptyViewHeaderHeight/index.ts
@@ -1,4 +1,4 @@
-import {BUTTON_HEIGHT, BUTTON_MARGIN, HEADER_HEIGHT} from './CONST';
+import {BUTTON_HEIGHT, BUTTON_MARGIN, HEADER_HEIGHT} from './const';
 
 function useEmptyViewHeaderHeight(isSmallScreenWidth: boolean): number {
     return isSmallScreenWidth ? HEADER_HEIGHT + BUTTON_HEIGHT + BUTTON_MARGIN : HEADER_HEIGHT;
